+++ hexo/package.json
@@ -3,7 +3,7 @@
   "version": "0.0.1",
   "private": true,
   "hexo": {
-    "version": "3.1.1"
+    "version": "3.2.0"
   },
   "dependencies": {
     "hexo": "^3.1.1",
