+++ package.json
@@ -49,6 +49,7 @@
     "gulp-symlink": "^2.1.3",
     "gulp-uglify": "^1.4.2",
     "gulp-uglifycss": "^1.0.4",
+    "jasmine":"^2.4",
     "matchdep": "^1.0.0",
     "mkpath": "^1.0.0",
     "run-sequence": "^1.1.4",
@@ -56,7 +57,7 @@
     "stylestats": "^6.0.0"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "./node_modules/.bin/jasmine"
   },
   "repository": {
     "type": "git",
